requirements.txt
alembic==1.13.1
anyio==4.4.0
APIFlask==2.1.1
apispec==6.6.1
bcrypt==4.1.2
black==24.10.0
blinker==1.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
marshmallow==3.21.1
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
onesignal-sdk==2.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.3.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.2
webargs==8.4.0
Werkzeug==3.0.2
zipp==3.18.1
zope.event==5.0
zope.interface==6.3
